requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
alembic==1.13.3
watchdog==5.0.3
gunicorn==22.0.0
pydantic==2.8.2
pytest==8.3.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
Flask-Cors==4.0.1

